ci: gate the whole Rust workspace on clippy, and clear draco-core
The clippy step covered only draco-io and draco-gltf, so draco-core and
draco-cpp-test-bridge had accumulated 25 warnings that nothing would ever
fail on. None was a correctness defect: same-type casts, loop variables used
only to index, redundant imports, a borrow that is immediately dereferenced,
and one leftover `"FAIL";` statement whose value was already produced by the
following expression.

Most were applied by `clippy --fix`; the five it cannot fix were done by hand.
Two of those touch draco-core `src` and are pure code moves -- `items after a
test module` means production code sat below `#[cfg(test)] mod tests`, and
clippy relocates it. That is 171 moved lines in the tex-coord prediction
scheme, which is parity-critical, so it was checked rather than trusted: the
multiset of added lines equals the multiset of removed ones in both files, so
nothing but order changed. All 84 test suites pass, including the C++ parity
and byte-comparison ones.

Widening the gate to `--workspace` needed one more fix that is invisible
locally. `draco-cpp-test-bridge` compiles against a prebuilt C++ Draco when it
is present and falls back to `cpp_test_bridge_disabled` stubs when it is not,
and CI has no such build -- so CI lints a different set of code than a
developer machine does. Simulating that configuration surfaced two errors in a
stub that mirrors an `extern "C"` signature, where the real declaration is not
linted at all. Both configurations are now clean, which is what the widened
gate actually promises.

+++ crates/draco-core/tests/compat_encoder_options.rs
@@ -62,7 +62,7 @@ fn create_mesh_with_attributes() -> Mesh {
         3,
         DataType::Float32,
         false,
-        num_points as usize,
+        num_points,
     );
 
     for y in 0..grid_size {
@@ -73,7 +73,7 @@ fn create_mesh_with_attributes() -> Mesh {
             // Use same sinusoidal variation as compat_encoding_speed.rs (0.2 not 0.3)
             let pz = (x as f32 * 0.2).sin() * (y as f32 * 0.2).cos() * 2.0;
 
-            let offset = index as usize * 3 * 4;
+            let offset = index * 3 * 4;
             pos_attr
                 .buffer_mut()
                 .update(&px.to_le_bytes(), Some(offset));
